more flexible than bbappend
Currently the dtappend files and logic can only be accessed when using
either linux-yocto or linux-yocto-dev. Modify how this mechanism works
so that the core is included in an *.inc file. This feature will
continue to be picked up automatically by linux-yocto and
linux-yocto-dev, but is also available for custom linux kernel recipes
by simply adding the following:

	require recipes-kernel/linux/linux-dtappend.inc

Signed-off-by: Trevor Woerner <twoerner@gmail.com>